image-automation-toolkit/requirements.txt
Pillow>=9.0.0
img2pdf>=0.4.0